renamed wflow.py -> wflow_sbm.py
