+++ hydromt_wflow/__init__.py
@@ -1,7 +1,7 @@
 """HydroMT plugin for wflow models."""
 
 from .version import __version__
-from .wflow import WflowModel
+from .wflow_sbm import WflowModel
 from .wflow_sediment import WflowSedimentModel
 
 __all__ = ["WflowModel", "WflowSedimentModel"]
