+++ tests/components/test_forcing_component.py
@@ -5,7 +5,7 @@ import pytest
 import xarray as xr
 
 from hydromt_wflow.components import WflowForcingComponent
-from hydromt_wflow.wflow import WflowModel
+from hydromt_wflow.wflow_sbm import WflowModel
 
 
 @pytest.fixture
